+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Also add the possiblity to disable advancing daily preps and checking for random encounters when resting
+* Disabling watch in rest popup does not disable random encounter checks anymore
 * Do not show rest config popup when clicking the "Continue" button
 
 ## [2.1.0] - 2025-01-26